Development work in Tromsø (Agoston, Veronika, Csilla, Trond): The old rules were made for a version of the lexc output that already was removed from the source file. We thus commented out those rules (at the end of the file, marked with !TOS).
Instead we made a rule set along the following philosophy:
- Only one contlex for the nouns
- Default suffix being -ыC, and default stem with floating vow
- Delete ы suff vowel after V stem, and ы floating vowel
  before suffixes containing vowels.
- Delete ь sign in stems before palatal vowels
- Insert й before -ыг suffix for и/ы stems
- Delete т (in Px3) in consonant stems

We also gave all nouns the same contlex _N01, and the
lemma as stem.

This should cover the lion's share of nominal morphophonology
(finegrained details to be tested after yaml files have been corrected)

We did not look at verbs at all, but much of this should hold for
verbs also.