Update mns-propernouns.lexc
No reason to write propernouns a second time, e.g., Саранпа̄выл, То̄рум were already present. The Sem/... tags, of course, need to be corrected.
Remember Russian like surnames can be temporarily added here, but they need to be moved to the URJ proper names if they are names that show up in other languages as well 100%.